Refactoring: Moved tests for filters under tests/filters
This is preparation work before making the actual changes to filtering and permissions.
